+++ changelog.md
@@ -6,7 +6,6 @@
 - Substantially reduced frequency of Forkbomb projectiles getting stuck inside walls
 - Bubblebeam's projectiles no longer collide with other projectiles belonging to the player until they have unbubbled something
 - Fixed Stereoscope's projectiles throwing an exception every frame when not fired from Stereoscope directly
-- Fixed Stereoscope's projectiles not working properly with Helix Bullets
 - Fixed Mastered Macheening's projectiles colliding with and destroying other player-owned projectiles
 - Fixed Empath's projectiles colliding with and destroying other player-owned projectiles
 - Fixed null deref caused by Suncaster projectiles not spawned by Suncaster trying to refract to their non-existent gun's prisms
@@ -20,6 +19,7 @@
 - Fixed noisy audio glitch caused by repeatedly taking 0 damage hits while under Sunderbuss or Macheening's curse
 - Allowed Wayfarer to have multiple projectiles under its control at the same time (e.g., if firing with Scattershot or Helix Bullets)
 - Made Wayfarer projectiles created without a controlling Wayfarer (e.g., via Chance Bullets) die on impact
+- Fixed Helix Bullets rendering Stereoscope, Sunderbuss, and Sextant unusable
 
 ## 1.26.0 (2025-05-29)